Added UIController script, which is a state machine that fire events that UIElement and its children listen for.
Replaced Menu with UIElement, an abstract class
which listens for state updates from the new UIController.

Heavlily simplifed GameManager, so it now primarily only handles setting
the scene and UIState.

Renamed MatchEndMenu to ResultsMenu.

Mainmenu, PauseMenu, ResultsMenu, and Scoreboard now inherit from
UIElement, and as such are automatically enabled and disabled when a
valid UI state is detected.

PauseMenu and ResultsMenu now freeze and unfreeze time themselves when
entering and exiting a valid state, respectively.

Added UIState, which holds a simple enum of various UI states.

Updated Singleton to only add top-level gameObjects to DontDestroyOnLoad
as I wanted my UI singletons to sit under the UIController on the
hierarchy, which is not possible if they are set to not destroy on load.
UIController is itself set to not be destroyed on load, so its children
are safe anyway.